requirements.txt
appdirs==1.4.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
email-validator==1.1.2
Flask==1.1.2
flask-mongoengine==1.0.0
Flask-WTF==0.14.3
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
lockfile==0.12.2
MarkupSafe==1.1.1
mongoengine==0.23.0
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
pymongo==3.11.3
pyparsing==2.4.6
python-dotenv==0.17.0
pytoml==0.1.21
requests==2.22.0
retrying==1.3.3
six==1.14.0
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
WTForms==2.3.3
